# src/api/swagger.yaml
swagger: '2.0'
host: 'deploy.dev.maaii.com:4004'
info:
  version: '1.2.0'
  title: IAM (Identity & Access Management Service)
  description: >
    IAM stores user and company information, provides OpenID Connect authentication.
paths:
...
  /access/roles:
    get:
      summary: Get roles
      operationId: getRoles
      tags:
        - Role
      produces:
        - application/json
      parameters:
        - name: company
          type: string
          in: query
          required: false
        - name: service
          type: string
          in: query
          required: false
      responses:
        '200':
            description: list of roles
            schema:
              type: array
              items:
                $ref: '#/definitions/RoleWithMeta'
        default:
          description: other error
          schema:
            $ref: '#/definitions/ErrorWrapper'
    post:
      summary: Create role
      operationId: createRole
      tags:
        - Role
      produces:
        - application/json
      parameters:
        - name: role
          in: body
          required: true
          schema:
            $ref: '#/definitions/Role'
      responses:
        '201':
          description: role was created
          schema:
            $ref: '#/definitions/RoleWithMeta'
        '422':
          description: validation error
          schema:
            $ref: '#/definitions/ErrorWrapper'
        default:
          description: other error
          schema:
            $ref: '#/definitions/ErrorWrapper'
...
  /access/users/{userId}/roles:
    get:
      summary: Get user roles
      operationId: getUserRoles
      tags:
        - Role
      produces:
        - application/json
      parameters:
        - name: userId
          type: string
          in: path
          required: true
        - name: service
          type: string
          in: query
          required: false
        - name: company
          type: string
          in: query
          required: false
      responses:
        '200':
            description: list of roles
            schema:
              type: array
              items:
                $ref: '#/definitions/RoleWithMeta'
        '404':
          description: not found error
          schema:
            $ref: '#/definitions/ErrorWrapper'
        default:
          description: other error
          schema:
            $ref: '#/definitions/ErrorWrapper'
    put:
      summary: Set user roles
      operationId: setUserRoles
      tags:
        - Role
      produces:
        - application/json
      parameters:
        - name: userId
          type: string
          in: path
          required: true
        - name: roles
          in: body
          required: true
          schema:
            type: array
            items:
              $ref: '#/definitions/RoleWithMeta'
      responses:
        '200':
            description: list of roles
            schema:
              type: array
              items:
                $ref: '#/definitions/RoleIdentity'
        '404':
          description: not found error
          schema:
            $ref: '#/definitions/ErrorWrapper'
        '422':
          description: validation error
          schema:
            $ref: '#/definitions/ErrorWrapper'
        default:
          description: other error
          schema:
            $ref: '#/definitions/ErrorWrapper'
  /access/users/{userId}/permissions:
    get:
      summary: Get user permissions
      operationId: getUserPermissions
      tags:
        - Role
      produces:
        - application/json
      parameters:
        - name: userId
          type: string
          in: path
          required: true
      responses:
        '200':
          description: permissions map (resource name -> array of actions)
          schema:
            type: object
        '404':
          description: not found error
          schema:
            $ref: '#/definitions/ErrorWrapper'
        '422':
          description: validation error
          schema:
            $ref: '#/definitions/ErrorWrapper'
        default:
          description: other error
          schema:
            $ref: '#/definitions/ErrorWrapper'
definitions:
  RoleWithMeta:
    type: object
    allOf:
      - $ref: '#/definitions/Role'
      - $ref: '#/definitions/RoleIdentity'
      - $ref: '#/definitions/TimestampMeta'
  RoleIdentity:
    type: object
    required:
      - id
    properties:
      id:
        type: string
  Role:
    type: object
    required:
      - name
      - company
      - service
    properties:
      isRoot:
        type: boolean
        description: root role cannot be deleted
      name:
        type: string
        description: role display name
      company:
        type: string
        description: company id
      service:
        type: string
      permissions:
        type: object
...
  ErrorWrapper:
    type: object
    required:
      - error
    properties:
      error:
        $ref: '#/definitions/Error'
  Error:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: number
        format: int32
        enum:
          - 20000
          - 20001
          - 20002
          - 20003
          - 20004
          - 20005
          - 20006
      message:
        type: string
...
  TimestampMeta:
    type: object
    properties:
      createdAt:
        type: string
        format: 'date-time'
      updatedAt:
        type: string
        format: 'date-time'
...